# api-documentation/configuration-api-swagger.yaml
openapi: 3.0.3
info:
  title: Smart Home Configuration API
  description: |
    API для конфигурации устройств умного дома.
    Предоставляет функциональность для регистрации устройств, управления конфигурацией и контроля доступа.
  version: 1.0.0
  contact:
    name: Smart Home Team
    email: support@smarthome.com

servers:
  - url: http://localhost:8080/api/v1
    description: Development server

tags:
  - name: devices
    description: Управление устройствами
  - name: configurations
    description: Конфигурация устройств  
  - name: locations
    description: Управление локациями (домами)

paths:
  # Device Management Endpoints
  /devices:
    get:
      tags:
        - devices
      summary: Получить список устройств
      description: Возвращает список всех зарегистрированных устройств
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: Список устройств успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - devices
      summary: Зарегистрировать новое устройство
      description: Регистрирует новое устройство в системе
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Устройство успешно зарегистрировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Device Configuration Endpoints
  /devices/{device_id}/configurations:
    get:
      tags:
        - configurations
      summary: Получить конфигурации устройства
      description: Возвращает все конфигурации для указанного устройства
      parameters:
        - name: device_id
          in: path
          required: true
          description: Уникальный идентификатор устройства
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: Конфигурации получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceConfiguration'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - configurations
      summary: Добавить конфигурацию устройства
      description: Добавляет новую конфигурацию для устройства
      parameters:
        - name: device_id
          in: path
          required: true
          description: Уникальный идентификатор устройства
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConfigurationRequest'
      responses:
        '201':
          description: Конфигурация создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceConfiguration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Location Management Endpoints
  /locations:
    get:
      tags:
        - locations
      summary: Получить список локаций
      description: Возвращает список всех локаций (домов)
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: Список локаций получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - locations
      summary: Создать новую локацию
      description: Создает новую локацию (дом)
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Локация создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    UserIdHeader:
      name: User-ID
      in: header
      required: true
      description: Уникальный идентификатор пользователя
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"

  schemas:
    Device:
      type: object
      required:
        - device_id
        - location_id
        - serial_id
        - name
        - type
        - status
        - created_at
        - updated_at
      properties:
        device_id:
          type: string
          format: uuid
          description: Уникальный идентификатор устройства
        location_id:
          type: string
          format: uuid
          description: Идентификатор локации, где установлено устройство
        serial_id:
          type: string
          maxLength: 255
          description: Серийный номер устройства (уникальный идентификатор оборудования)
          example: "SN123456789"
        name:
          type: string
          maxLength: 255
          description: Название устройства
          example: "Термостат в гостиной"
        type:
          type: string
          maxLength: 100
          description: Тип устройства
          example: "thermostat"
        status:
          type: string
          enum: [active, inactive, error, maintenance]
          description: Текущий статус устройства
        created_at:
          type: string
          format: date-time
          description: Время создания записи
        updated_at:
          type: string
          format: date-time
          description: Время последнего обновления

    CreateDeviceRequest:
      type: object
      required:
        - location_id
        - serial_id
        - name
        - type
      properties:
        location_id:
          type: string
          format: uuid
          description: Идентификатор локации для устройства
        serial_id:
          type: string
          maxLength: 255
          description: Серийный номер устройства (уникальный идентификатор оборудования)
          example: "SN123456789"
        name:
          type: string
          maxLength: 255
          description: Название устройства
          example: "Термостат в гостиной"
        type:
          type: string
          maxLength: 100
          description: Тип устройства
          example: "thermostat"
        status:
          type: string
          enum: [active, inactive, error, maintenance]
          description: Начальный статус устройства
          default: "active"



    DeviceConfiguration:
      type: object
      required:
        - config_id
        - device_id
        - config_name
        - config_value
        - config_type
        - is_active
        - created_at
        - updated_at
      properties:
        config_id:
          type: string
          format: uuid
          description: Уникальный идентификатор конфигурации
        device_id:
          type: string
          format: uuid
          description: Идентификатор устройства
        config_name:
          type: string
          maxLength: 255
          description: Название конфигурационного параметра
          example: "temperature_threshold"
        config_value:
          type: string
          description: Значение конфигурационного параметра
          example: "25.5"
        config_type:
          type: string
          maxLength: 50
          description: Тип конфигурационного параметра
          enum: [string, number, boolean, json]
          example: "number"
        is_active:
          type: boolean
          description: Активна ли данная конфигурация
        created_at:
          type: string
          format: date-time
          description: Время создания конфигурации
        updated_at:
          type: string
          format: date-time
          description: Время последнего обновления конфигурации

    CreateConfigurationRequest:
      type: object
      required:
        - config_name
        - config_value
        - config_type
      properties:
        config_name:
          type: string
          maxLength: 255
          description: Название конфигурационного параметра
          example: "temperature_threshold"
        config_value:
          type: string
          description: Значение конфигурационного параметра
          example: "25.5"
        config_type:
          type: string
          maxLength: 50
          description: Тип конфигурационного параметра
          enum: [string, number, boolean, json]
          example: "number"
        is_active:
          type: boolean
          description: Активна ли конфигурация
          default: true



    Location:
      type: object
      required:
        - location_id
        - name
      properties:
        location_id:
          type: string
          format: uuid
          description: Уникальный идентификатор локации
        name:
          type: string
          maxLength: 255
          description: Название локации
          example: "Мой дом"

    CreateLocationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
          description: Название локации
          example: "Мой дом"



    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Код ошибки
          example: "DEVICE_NOT_FOUND"
        message:
          type: string
          description: Описание ошибки
          example: "Устройство с указанным ID не найдено"
        details:
          type: object
          description: Дополнительные детали ошибки

  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INVALID_REQUEST"
            message: "Переданы некорректные параметры"

    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NOT_FOUND"
            message: "Запрашиваемый ресурс не найден"

    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INTERNAL_ERROR"
            message: "Произошла внутренняя ошибка сервера"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: [] 